@@ -22,5 +22,5 @@ flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
   flutter: ">=1.9.1+hotfix.4 <2.0.0"
