@@ -1,3 +1,7 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 package io.flutter.plugins.camera;
 
 import static android.view.OrientationEventListener.ORIENTATION_UNKNOWN;
@@ -33,18 +37,21 @@ import android.os.Build.VERSION;
 import android.os.Build.VERSION_CODES;
 import android.os.Handler;
 import android.os.Looper;
+import android.util.Log;
 import android.util.Range;
 import android.util.Rational;
 import android.util.Size;
 import android.view.OrientationEventListener;
 import android.view.Surface;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import io.flutter.plugin.common.EventChannel;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugins.camera.PictureCaptureRequest.State;
 import io.flutter.plugins.camera.media.MediaRecorderBuilder;
 import io.flutter.plugins.camera.types.ExposureMode;
 import io.flutter.plugins.camera.types.FlashMode;
+import io.flutter.plugins.camera.types.FocusMode;
 import io.flutter.plugins.camera.types.ResolutionPreset;
 import io.flutter.view.TextureRegistry.SurfaceTextureEntry;
 import java.io.File;
@@ -59,7 +66,14 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.Executors;
 
+@FunctionalInterface
+interface ErrorCallback {
+  void onError(String errorCode, String errorMessage);
+}
+
 public class Camera {
+  private static final String TAG = "Camera";
+
   private final SurfaceTextureEntry flutterTexture;
   private final CameraManager cameraManager;
   private final OrientationEventListener orientationEventListener;
@@ -73,6 +87,7 @@ public class Camera {
   private final CamcorderProfile recordingProfile;
   private final DartMessenger dartMessenger;
   private final CameraZoom cameraZoom;
+  private final CameraCharacteristics cameraCharacteristics;
 
   private CameraDevice cameraDevice;
   private CameraCaptureSession cameraCaptureSession;
@@ -85,9 +100,20 @@ public class Camera {
   private int currentOrientation = ORIENTATION_UNKNOWN;
   private FlashMode flashMode;
   private ExposureMode exposureMode;
+  private FocusMode focusMode;
   private PictureCaptureRequest pictureCaptureRequest;
   private CameraRegions cameraRegions;
   private int exposureOffset;
+  private boolean useAutoFocus = true;
+  private Range<Integer> fpsRange;
+
+  private static final HashMap<String, Integer> supportedImageFormats;
+  // Current supported outputs
+  static {
+    supportedImageFormats = new HashMap<>();
+    supportedImageFormats.put("yuv420", 35);
+    supportedImageFormats.put("jpeg", 256);
+  }
 
   public Camera(
       final Activity activity,
@@ -108,6 +134,7 @@ public class Camera {
     this.applicationContext = activity.getApplicationContext();
     this.flashMode = FlashMode.auto;
     this.exposureMode = ExposureMode.auto;
+    this.focusMode = FocusMode.auto;
     this.exposureOffset = 0;
     orientationEventListener =
         new OrientationEventListener(activity.getApplicationContext()) {
@@ -122,10 +149,12 @@ public class Camera {
         };
     orientationEventListener.enable();
 
-    CameraCharacteristics characteristics = cameraManager.getCameraCharacteristics(cameraName);
-    sensorOrientation = characteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
+    cameraCharacteristics = cameraManager.getCameraCharacteristics(cameraName);
+    initFps(cameraCharacteristics);
+    sensorOrientation = cameraCharacteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
     isFrontFacing =
-        characteristics.get(CameraCharacteristics.LENS_FACING) == CameraMetadata.LENS_FACING_FRONT;
+        cameraCharacteristics.get(CameraCharacteristics.LENS_FACING)
+            == CameraMetadata.LENS_FACING_FRONT;
     ResolutionPreset preset = ResolutionPreset.valueOf(resolutionPreset);
     recordingProfile =
         CameraUtils.getBestAvailableCamcorderProfileForResolutionPreset(cameraName, preset);
@@ -133,8 +162,29 @@ public class Camera {
     previewSize = computeBestPreviewSize(cameraName, preset);
     cameraZoom =
         new CameraZoom(
-            characteristics.get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE),
-            characteristics.get(CameraCharacteristics.SCALER_AVAILABLE_MAX_DIGITAL_ZOOM));
+            cameraCharacteristics.get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE),
+            cameraCharacteristics.get(CameraCharacteristics.SCALER_AVAILABLE_MAX_DIGITAL_ZOOM));
+  }
+
+  private void initFps(CameraCharacteristics cameraCharacteristics) {
+    try {
+      Range<Integer>[] ranges =
+          cameraCharacteristics.get(CameraCharacteristics.CONTROL_AE_AVAILABLE_TARGET_FPS_RANGES);
+      if (ranges != null) {
+        for (Range<Integer> range : ranges) {
+          int upper = range.getUpper();
+          Log.i("Camera", "[FPS Range Available] is:" + range);
+          if (upper >= 10) {
+            if (fpsRange == null || upper > fpsRange.getUpper()) {
+              fpsRange = range;
+            }
+          }
+        }
+      }
+    } catch (Exception e) {
+      e.printStackTrace();
+    }
+    Log.i("Camera", "[FPS Range] is:" + fpsRange);
   }
 
   private void prepareMediaRecorder(String outputFilePath) throws IOException {
@@ -150,15 +200,20 @@ public class Camera {
   }
 
   @SuppressLint("MissingPermission")
-  public void open() throws CameraAccessException {
+  public void open(String imageFormatGroup) throws CameraAccessException {
     pictureImageReader =
         ImageReader.newInstance(
             captureSize.getWidth(), captureSize.getHeight(), ImageFormat.JPEG, 2);
 
+    Integer imageFormat = supportedImageFormats.get(imageFormatGroup);
+    if (imageFormat == null) {
+      Log.w(TAG, "The selected imageFormatGroup is not supported by Android. Defaulting to yuv420");
+      imageFormat = ImageFormat.YUV_420_888;
+    }
+
     // Used to steam image byte data to dart side.
     imageStreamReader =
-        ImageReader.newInstance(
-            previewSize.getWidth(), previewSize.getHeight(), ImageFormat.YUV_420_888, 2);
+        ImageReader.newInstance(previewSize.getWidth(), previewSize.getHeight(), imageFormat, 2);
 
     cameraManager.openCamera(
         cameraName,
@@ -173,7 +228,9 @@ public class Camera {
                   previewSize.getWidth(),
                   previewSize.getHeight(),
                   exposureMode,
-                  isExposurePointSupported());
+                  focusMode,
+                  isExposurePointSupported(),
+                  isFocusPointSupported());
             } catch (CameraAccessException e) {
               dartMessenger.sendCameraErrorEvent(e.getMessage());
               close();
@@ -221,6 +278,118 @@ public class Camera {
         null);
   }
 
+  private void createCaptureSession(int templateType, Surface... surfaces)
+      throws CameraAccessException {
+    createCaptureSession(templateType, null, surfaces);
+  }
+
+  private void createCaptureSession(
+      int templateType, Runnable onSuccessCallback, Surface... surfaces)
+      throws CameraAccessException {
+    // Close any existing capture session.
+    closeCaptureSession();
+
+    // Create a new capture builder.
+    captureRequestBuilder = cameraDevice.createCaptureRequest(templateType);
+
+    // Build Flutter surface to render to
+    SurfaceTexture surfaceTexture = flutterTexture.surfaceTexture();
+    surfaceTexture.setDefaultBufferSize(previewSize.getWidth(), previewSize.getHeight());
+    Surface flutterSurface = new Surface(surfaceTexture);
+    captureRequestBuilder.addTarget(flutterSurface);
+
+    List<Surface> remainingSurfaces = Arrays.asList(surfaces);
+    if (templateType != CameraDevice.TEMPLATE_PREVIEW) {
+      // If it is not preview mode, add all surfaces as targets.
+      for (Surface surface : remainingSurfaces) {
+        captureRequestBuilder.addTarget(surface);
+      }
+    }
+
+    // Prepare the callback
+    CameraCaptureSession.StateCallback callback =
+        new CameraCaptureSession.StateCallback() {
+          @Override
+          public void onConfigured(@NonNull CameraCaptureSession session) {
+            if (cameraDevice == null) {
+              dartMessenger.sendCameraErrorEvent("The camera was closed during configuration.");
+              return;
+            }
+            cameraCaptureSession = session;
+
+            updateFpsRange();
+            updateFocus(focusMode);
+            updateFlash(flashMode);
+            updateExposure(exposureMode);
+
+            refreshPreviewCaptureSession(
+                onSuccessCallback, (code, message) -> dartMessenger.sendCameraErrorEvent(message));
+          }
+
+          @Override
+          public void onConfigureFailed(@NonNull CameraCaptureSession cameraCaptureSession) {
+            dartMessenger.sendCameraErrorEvent("Failed to configure camera session.");
+          }
+        };
+
+    // Start the session
+    if (VERSION.SDK_INT >= VERSION_CODES.P) {
+      // Collect all surfaces we want to render to.
+      List<OutputConfiguration> configs = new ArrayList<>();
+      configs.add(new OutputConfiguration(flutterSurface));
+      for (Surface surface : remainingSurfaces) {
+        configs.add(new OutputConfiguration(surface));
+      }
+      createCaptureSessionWithSessionConfig(configs, callback);
+    } else {
+      // Collect all surfaces we want to render to.
+      List<Surface> surfaceList = new ArrayList<>();
+      surfaceList.add(flutterSurface);
+      surfaceList.addAll(remainingSurfaces);
+      createCaptureSession(surfaceList, callback);
+    }
+  }
+
+  @TargetApi(VERSION_CODES.P)
+  private void createCaptureSessionWithSessionConfig(
+      List<OutputConfiguration> outputConfigs, CameraCaptureSession.StateCallback callback)
+      throws CameraAccessException {
+    cameraDevice.createCaptureSession(
+        new SessionConfiguration(
+            SessionConfiguration.SESSION_REGULAR,
+            outputConfigs,
+            Executors.newSingleThreadExecutor(),
+            callback));
+  }
+
+  @TargetApi(VERSION_CODES.LOLLIPOP)
+  @SuppressWarnings("deprecation")
+  private void createCaptureSession(
+      List<Surface> surfaces, CameraCaptureSession.StateCallback callback)
+      throws CameraAccessException {
+    cameraDevice.createCaptureSession(surfaces, callback, null);
+  }
+
+  private void refreshPreviewCaptureSession(
+      @Nullable Runnable onSuccessCallback, @NonNull ErrorCallback onErrorCallback) {
+    if (cameraCaptureSession == null) {
+      return;
+    }
+
+    try {
+      cameraCaptureSession.setRepeatingRequest(
+          captureRequestBuilder.build(),
+          pictureCaptureCallback,
+          new Handler(Looper.getMainLooper()));
+
+      if (onSuccessCallback != null) {
+        onSuccessCallback.run();
+      }
+    } catch (CameraAccessException | IllegalStateException | IllegalArgumentException e) {
+      onErrorCallback.onError("cameraAccess", e.getMessage());
+    }
+  }
+
   private void writeToFile(ByteBuffer buffer, File file) throws IOException {
     try (FileOutputStream outputStream = new FileOutputStream(file)) {
       while (0 < buffer.remaining()) {
@@ -261,7 +430,11 @@ public class Camera {
         },
         null);
 
-    runPictureAutoFocus();
+    if (useAutoFocus) {
+      runPictureAutoFocus();
+    } else {
+      runPicturePreCapture();
+    }
   }
 
   private final CameraCaptureSession.CaptureCallback pictureCaptureCallback =
@@ -344,8 +517,9 @@ public class Camera {
 
   private void runPictureAutoFocus() {
     assert (pictureCaptureRequest != null);
+
     pictureCaptureRequest.setState(PictureCaptureRequest.State.focusing);
-    lockAutoFocus();
+    lockAutoFocus(pictureCaptureCallback);
   }
 
   private void runPicturePreCapture() {
@@ -355,14 +529,13 @@ public class Camera {
     captureRequestBuilder.set(
         CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
         CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_START);
-    try {
-      cameraCaptureSession.capture(captureRequestBuilder.build(), pictureCaptureCallback, null);
-      captureRequestBuilder.set(
-          CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
-          CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_IDLE);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
+
+    refreshPreviewCaptureSession(
+        () ->
+            captureRequestBuilder.set(
+                CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
+                CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_IDLE),
+        (code, message) -> pictureCaptureRequest.error(code, message, null));
   }
 
   private void runPictureCapture() {
@@ -406,128 +579,28 @@ public class Camera {
     }
   }
 
-  private void lockAutoFocus() {
+  private void lockAutoFocus(CaptureCallback callback) {
     captureRequestBuilder.set(
         CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_START);
-    try {
-      cameraCaptureSession.capture(captureRequestBuilder.build(), pictureCaptureCallback, null);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
+
+    refreshPreviewCaptureSession(
+        null, (code, message) -> pictureCaptureRequest.error(code, message, null));
   }
 
   private void unlockAutoFocus() {
     captureRequestBuilder.set(
         CaptureRequest.CONTROL_AF_TRIGGER, CameraMetadata.CONTROL_AF_TRIGGER_CANCEL);
-    initPreviewCaptureBuilder();
+    updateFocus(focusMode);
     try {
       cameraCaptureSession.capture(captureRequestBuilder.build(), null, null);
     } catch (CameraAccessException ignored) {
     }
     captureRequestBuilder.set(
         CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_IDLE);
-    try {
-      cameraCaptureSession.setRepeatingRequest(
-          captureRequestBuilder.build(), pictureCaptureCallback, null);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
-  }
 
-  private void createCaptureSession(int templateType, Surface... surfaces)
-      throws CameraAccessException {
-    createCaptureSession(templateType, null, surfaces);
-  }
-
-  private void createCaptureSession(
-      int templateType, Runnable onSuccessCallback, Surface... surfaces)
-      throws CameraAccessException {
-    // Close any existing capture session.
-    closeCaptureSession();
-
-    // Create a new capture builder.
-    captureRequestBuilder = cameraDevice.createCaptureRequest(templateType);
-
-    // Build Flutter surface to render to
-    SurfaceTexture surfaceTexture = flutterTexture.surfaceTexture();
-    surfaceTexture.setDefaultBufferSize(previewSize.getWidth(), previewSize.getHeight());
-    Surface flutterSurface = new Surface(surfaceTexture);
-    captureRequestBuilder.addTarget(flutterSurface);
-
-    List<Surface> remainingSurfaces = Arrays.asList(surfaces);
-    if (templateType != CameraDevice.TEMPLATE_PREVIEW) {
-      // If it is not preview mode, add all surfaces as targets.
-      for (Surface surface : remainingSurfaces) {
-        captureRequestBuilder.addTarget(surface);
-      }
-    }
-
-    // Prepare the callback
-    CameraCaptureSession.StateCallback callback =
-        new CameraCaptureSession.StateCallback() {
-          @Override
-          public void onConfigured(@NonNull CameraCaptureSession session) {
-            try {
-              if (cameraDevice == null) {
-                dartMessenger.sendCameraErrorEvent("The camera was closed during configuration.");
-                return;
-              }
-              cameraCaptureSession = session;
-              initPreviewCaptureBuilder();
-              cameraCaptureSession.setRepeatingRequest(
-                  captureRequestBuilder.build(),
-                  pictureCaptureCallback,
-                  new Handler(Looper.getMainLooper()));
-              if (onSuccessCallback != null) {
-                onSuccessCallback.run();
-              }
-            } catch (CameraAccessException | IllegalStateException | IllegalArgumentException e) {
-              dartMessenger.sendCameraErrorEvent(e.getMessage());
-            }
-          }
-
-          @Override
-          public void onConfigureFailed(@NonNull CameraCaptureSession cameraCaptureSession) {
-            dartMessenger.sendCameraErrorEvent("Failed to configure camera session.");
-          }
-        };
-
-    // Start the session
-    if (VERSION.SDK_INT >= VERSION_CODES.P) {
-      // Collect all surfaces we want to render to.
-      List<OutputConfiguration> configs = new ArrayList<>();
-      configs.add(new OutputConfiguration(flutterSurface));
-      for (Surface surface : remainingSurfaces) {
-        configs.add(new OutputConfiguration(surface));
-      }
-      createCaptureSessionWithSessionConfig(configs, callback);
-    } else {
-      // Collect all surfaces we want to render to.
-      List<Surface> surfaceList = new ArrayList<>();
-      surfaceList.add(flutterSurface);
-      surfaceList.addAll(remainingSurfaces);
-      createCaptureSession(surfaceList, callback);
-    }
-  }
-
-  @TargetApi(VERSION_CODES.P)
-  private void createCaptureSessionWithSessionConfig(
-      List<OutputConfiguration> outputConfigs, CameraCaptureSession.StateCallback callback)
-      throws CameraAccessException {
-    cameraDevice.createCaptureSession(
-        new SessionConfiguration(
-            SessionConfiguration.SESSION_REGULAR,
-            outputConfigs,
-            Executors.newSingleThreadExecutor(),
-            callback));
-  }
-
-  @TargetApi(VERSION_CODES.LOLLIPOP)
-  @SuppressWarnings("deprecation")
-  private void createCaptureSession(
-      List<Surface> surfaces, CameraCaptureSession.StateCallback callback)
-      throws CameraAccessException {
-    cameraDevice.createCaptureSession(surfaces, callback, null);
+    refreshPreviewCaptureSession(
+        null,
+        (errorCode, errorMessage) -> pictureCaptureRequest.error(errorCode, errorMessage, null));
   }
 
   public void startVideoRecording(Result result) {
@@ -560,8 +633,14 @@ public class Camera {
 
     try {
       recordingVideo = false;
-      closeCaptureSession();
-      mediaRecorder.stop();
+
+      try {
+        cameraCaptureSession.abortCaptures();
+        mediaRecorder.stop();
+      } catch (CameraAccessException | IllegalStateException e) {
+        // Ignore exceptions and try to continue (changes are camera session already aborted capture)
+      }
+
       mediaRecorder.reset();
       startPreview();
       result.success(videoRecordingFile.getAbsolutePath());
@@ -630,8 +709,8 @@ public class Camera {
 
     // If switching directly from torch to auto or on, make sure we turn off the torch.
     if (flashMode == FlashMode.torch && mode != FlashMode.torch && mode != FlashMode.off) {
-      this.flashMode = FlashMode.off;
-      initPreviewCaptureBuilder();
+      updateFlash(FlashMode.off);
+
       this.cameraCaptureSession.setRepeatingRequest(
           captureRequestBuilder.build(),
           new CaptureCallback() {
@@ -647,8 +726,13 @@ public class Camera {
               }
 
               updateFlash(mode);
-              result.success(null);
-              isFinished = true;
+              refreshPreviewCaptureSession(
+                  () -> {
+                    result.success(null);
+                    isFinished = true;
+                  },
+                  (code, message) ->
+                      result.error("setFlashModeFailed", "Could not set flash mode.", null));
             }
 
             @Override
@@ -667,26 +751,16 @@ public class Camera {
           null);
     } else {
       updateFlash(mode);
-      result.success(null);
-    }
-  }
 
-  private void updateFlash(FlashMode mode) {
-    // Get flash
-    flashMode = mode;
-    initPreviewCaptureBuilder();
-    try {
-      cameraCaptureSession.setRepeatingRequest(
-          captureRequestBuilder.build(), pictureCaptureCallback, null);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
+      refreshPreviewCaptureSession(
+          () -> result.success(null),
+          (code, message) -> result.error("setFlashModeFailed", "Could not set flash mode.", null));
     }
   }
 
   public void setExposureMode(@NonNull final Result result, ExposureMode mode)
       throws CameraAccessException {
-    this.exposureMode = mode;
-    initPreviewCaptureBuilder();
+    updateExposure(mode);
     cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
     result.success(null);
   }
@@ -699,24 +773,70 @@ public class Camera {
           "setExposurePointFailed", "Device does not have exposure point capabilities", null);
       return;
     }
-    // Check if we are doing a reset or not
-    if (x == null || y == null) {
-      x = 0.5;
-      y = 0.5;
-    }
-    // Get the current region boundaries.
-    Size maxBoundaries = getRegionBoundaries();
-    if (maxBoundaries == null) {
+    // Check if the current region boundaries are known
+    if (cameraRegions.getMaxBoundaries() == null) {
       result.error("setExposurePointFailed", "Could not determine max region boundaries", null);
       return;
     }
     // Set the metering rectangle
-    cameraRegions.setAutoExposureMeteringRectangleFromPoint(x, y);
+    if (x == null || y == null) cameraRegions.resetAutoExposureMeteringRectangle();
+    else cameraRegions.setAutoExposureMeteringRectangleFromPoint(x, y);
     // Apply it
-    initPreviewCaptureBuilder();
-    this.cameraCaptureSession.setRepeatingRequest(
-        captureRequestBuilder.build(), pictureCaptureCallback, null);
+    updateExposure(exposureMode);
+    refreshPreviewCaptureSession(
+        () -> result.success(null), (code, message) -> result.error("CameraAccess", message, null));
+  }
+
+  public void setFocusMode(@NonNull final Result result, FocusMode mode)
+      throws CameraAccessException {
+    this.focusMode = mode;
+
+    updateFocus(mode);
+
+    switch (mode) {
+      case auto:
+        refreshPreviewCaptureSession(
+            null, (code, message) -> result.error("setFocusMode", message, null));
+        break;
+      case locked:
+        lockAutoFocus(
+            new CaptureCallback() {
+              @Override
+              public void onCaptureCompleted(
+                  @NonNull CameraCaptureSession session,
+                  @NonNull CaptureRequest request,
+                  @NonNull TotalCaptureResult result) {
+                unlockAutoFocus();
+              }
+            });
+        break;
+    }
     result.success(null);
+  }
+
+  public void setFocusPoint(@NonNull final Result result, Double x, Double y)
+      throws CameraAccessException {
+    // Check if focus point functionality is available.
+    if (!isFocusPointSupported()) {
+      result.error("setFocusPointFailed", "Device does not have focus point capabilities", null);
+      return;
+    }
+
+    // Check if the current region boundaries are known
+    if (cameraRegions.getMaxBoundaries() == null) {
+      result.error("setFocusPointFailed", "Could not determine max region boundaries", null);
+      return;
+    }
+
+    // Set the metering rectangle
+    if (x == null || y == null) {
+      cameraRegions.resetAutoFocusMeteringRectangle();
+    } else {
+      cameraRegions.setAutoFocusMeteringRectangleFromPoint(x, y);
+    }
+
+    // Apply the new metering rectangle
+    setFocusMode(result, focusMode);
   }
 
   @TargetApi(VERSION_CODES.P)
@@ -768,6 +888,14 @@ public class Camera {
     return supportedRegions != null && supportedRegions > 0;
   }
 
+  private boolean isFocusPointSupported() throws CameraAccessException {
+    Integer supportedRegions =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.CONTROL_MAX_REGIONS_AF);
+    return supportedRegions != null && supportedRegions > 0;
+  }
+
   public double getMinExposureOffset() throws CameraAccessException {
     Range<Integer> range =
         cameraManager
@@ -802,13 +930,114 @@ public class Camera {
     double stepSize = getExposureOffsetStepSize();
     exposureOffset = (int) (offset / stepSize);
     // Apply it
-    initPreviewCaptureBuilder();
+    updateExposure(exposureMode);
     this.cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
     result.success(offset);
   }
 
-  private void initPreviewCaptureBuilder() {
-    captureRequestBuilder.set(CaptureRequest.CONTROL_MODE, CaptureRequest.CONTROL_MODE_AUTO);
+  public float getMaxZoomLevel() {
+    return cameraZoom.maxZoom;
+  }
+
+  public float getMinZoomLevel() {
+    return CameraZoom.DEFAULT_ZOOM_FACTOR;
+  }
+
+  public void setZoomLevel(@NonNull final Result result, float zoom) throws CameraAccessException {
+    float maxZoom = cameraZoom.maxZoom;
+    float minZoom = CameraZoom.DEFAULT_ZOOM_FACTOR;
+
+    if (zoom > maxZoom || zoom < minZoom) {
+      String errorMessage =
+          String.format(
+              Locale.ENGLISH,
+              "Zoom level out of bounds (zoom level should be between %f and %f).",
+              minZoom,
+              maxZoom);
+      result.error("ZOOM_ERROR", errorMessage, null);
+      return;
+    }
+
+    //Zoom area is calculated relative to sensor area (activeRect)
+    if (captureRequestBuilder != null) {
+      final Rect computedZoom = cameraZoom.computeZoom(zoom);
+      captureRequestBuilder.set(CaptureRequest.SCALER_CROP_REGION, computedZoom);
+      cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
+    }
+
+    result.success(null);
+  }
+
+  private void updateFpsRange() {
+    if (fpsRange == null) {
+      return;
+    }
+
+    captureRequestBuilder.set(CaptureRequest.CONTROL_AE_TARGET_FPS_RANGE, fpsRange);
+  }
+
+  private void updateFocus(FocusMode mode) {
+    if (useAutoFocus) {
+      int[] modes = cameraCharacteristics.get(CameraCharacteristics.CONTROL_AF_AVAILABLE_MODES);
+      // Auto focus is not supported
+      if (modes == null
+          || modes.length == 0
+          || (modes.length == 1 && modes[0] == CameraCharacteristics.CONTROL_AF_MODE_OFF)) {
+        useAutoFocus = false;
+        captureRequestBuilder.set(
+            CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_OFF);
+      } else {
+        // Applying auto focus
+        switch (mode) {
+          case locked:
+            captureRequestBuilder.set(
+                CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_AUTO);
+            break;
+          case auto:
+            captureRequestBuilder.set(
+                CaptureRequest.CONTROL_AF_MODE,
+                recordingVideo
+                    ? CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_VIDEO
+                    : CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
+          default:
+            break;
+        }
+        MeteringRectangle afRect = cameraRegions.getAFMeteringRectangle();
+        captureRequestBuilder.set(
+            CaptureRequest.CONTROL_AF_REGIONS,
+            afRect == null ? null : new MeteringRectangle[] {afRect});
+      }
+    } else {
+      captureRequestBuilder.set(CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_OFF);
+    }
+  }
+
+  private void updateExposure(ExposureMode mode) {
+    exposureMode = mode;
+
+    // Applying auto exposure
+    MeteringRectangle aeRect = cameraRegions.getAEMeteringRectangle();
+    captureRequestBuilder.set(
+        CaptureRequest.CONTROL_AE_REGIONS,
+        aeRect == null ? null : new MeteringRectangle[] {cameraRegions.getAEMeteringRectangle()});
+
+    switch (mode) {
+      case locked:
+        captureRequestBuilder.set(CaptureRequest.CONTROL_AE_LOCK, true);
+        break;
+      case auto:
+      default:
+        captureRequestBuilder.set(CaptureRequest.CONTROL_AE_LOCK, false);
+        break;
+    }
+
+    captureRequestBuilder.set(CaptureRequest.CONTROL_AE_EXPOSURE_COMPENSATION, exposureOffset);
+  }
+
+  private void updateFlash(FlashMode mode) {
+    // Get flash
+    flashMode = mode;
+
     // Applying flash modes
     switch (flashMode) {
       case off:
@@ -833,24 +1062,6 @@ public class Camera {
         captureRequestBuilder.set(CaptureRequest.FLASH_MODE, CaptureRequest.FLASH_MODE_TORCH);
         break;
     }
-    // Applying auto exposure
-    MeteringRectangle aeRect = cameraRegions.getAEMeteringRectangle();
-    captureRequestBuilder.set(
-        CaptureRequest.CONTROL_AE_REGIONS,
-        aeRect == null ? null : new MeteringRectangle[] {cameraRegions.getAEMeteringRectangle()});
-    switch (exposureMode) {
-      case locked:
-        captureRequestBuilder.set(CaptureRequest.CONTROL_AE_LOCK, true);
-        break;
-      case auto:
-      default:
-        captureRequestBuilder.set(CaptureRequest.CONTROL_AE_LOCK, false);
-        break;
-    }
-    captureRequestBuilder.set(CaptureRequest.CONTROL_AE_EXPOSURE_COMPENSATION, exposureOffset);
-    // Applying auto focus
-    captureRequestBuilder.set(
-        CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
   }
 
   public void startPreview() throws CameraAccessException {
@@ -910,37 +1121,11 @@ public class Camera {
         null);
   }
 
-  public float getMaxZoomLevel() {
-    return cameraZoom.maxZoom;
-  }
-
-  public float getMinZoomLevel() {
-    return CameraZoom.DEFAULT_ZOOM_FACTOR;
-  }
-
-  public void setZoomLevel(@NonNull final Result result, float zoom) throws CameraAccessException {
-    float maxZoom = cameraZoom.maxZoom;
-    float minZoom = CameraZoom.DEFAULT_ZOOM_FACTOR;
-
-    if (zoom > maxZoom || zoom < minZoom) {
-      String errorMessage =
-          String.format(
-              Locale.ENGLISH,
-              "Zoom level out of bounds (zoom level should be between %f and %f).",
-              minZoom,
-              maxZoom);
-      result.error("ZOOM_ERROR", errorMessage, null);
-      return;
+  public void stopImageStream() throws CameraAccessException {
+    if (imageStreamReader != null) {
+      imageStreamReader.setOnImageAvailableListener(null, null);
     }
-
-    //Zoom area is calculated relative to sensor area (activeRect)
-    if (captureRequestBuilder != null) {
-      final Rect computedZoom = cameraZoom.computeZoom(zoom);
-      captureRequestBuilder.set(CaptureRequest.SCALER_CROP_REGION, computedZoom);
-      cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
-    }
-
-    result.success(null);
+    startPreview();
   }
 
   private void closeCaptureSession() {
