@@ -1,3 +1,7 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 package io.flutter.plugins.camera;
 
 import android.app.Activity;
@@ -12,6 +16,7 @@ import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugins.camera.CameraPermissions.PermissionsRegistry;
 import io.flutter.plugins.camera.types.ExposureMode;
 import io.flutter.plugins.camera.types.FlashMode;
+import io.flutter.plugins.camera.types.FocusMode;
 import io.flutter.view.TextureRegistry;
 import java.util.HashMap;
 import java.util.Map;
@@ -80,7 +85,7 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
         {
           if (camera != null) {
             try {
-              camera.open();
+              camera.open(call.argument("imageFormatGroup"));
               result.success(null);
             } catch (Exception e) {
               handleException(e, result);
@@ -206,6 +211,37 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
           }
           break;
         }
+      case "setFocusMode":
+        {
+          String modeStr = call.argument("mode");
+          FocusMode mode = FocusMode.getValueForString(modeStr);
+          if (mode == null) {
+            result.error("setFocusModeFailed", "Unknown focus mode " + modeStr, null);
+            return;
+          }
+          try {
+            camera.setFocusMode(result, mode);
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "setFocusPoint":
+        {
+          Boolean reset = call.argument("reset");
+          Double x = null;
+          Double y = null;
+          if (reset == null || !reset) {
+            x = call.argument("x");
+            y = call.argument("y");
+          }
+          try {
+            camera.setFocusPoint(result, x, y);
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
       case "startImageStream":
         {
           try {
@@ -219,7 +255,7 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
       case "stopImageStream":
         {
           try {
-            camera.startPreview();
+            camera.stopImageStream();
             result.success(null);
           } catch (Exception e) {
             handleException(e, result);
