@@ -1,3 +1,7 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 package io.flutter.plugins.camera;
 
 import static junit.framework.TestCase.assertNull;
@@ -8,6 +12,7 @@ import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.StandardMethodCodec;
 import io.flutter.plugins.camera.types.ExposureMode;
+import io.flutter.plugins.camera.types.FocusMode;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
@@ -59,7 +64,7 @@ public class DartMessengerTest {
 
   @Test
   public void sendCameraInitializedEvent_includesPreviewSize() {
-    dartMessenger.sendCameraInitializedEvent(0, 0, ExposureMode.auto, true);
+    dartMessenger.sendCameraInitializedEvent(0, 0, ExposureMode.auto, FocusMode.auto, true, true);
 
     List<ByteBuffer> sentMessages = fakeBinaryMessenger.getMessages();
     assertEquals(1, sentMessages.size());
@@ -68,7 +73,9 @@ public class DartMessengerTest {
     assertEquals(0, (double) call.argument("previewWidth"), 0);
     assertEquals(0, (double) call.argument("previewHeight"), 0);
     assertEquals("ExposureMode auto", call.argument("exposureMode"), "auto");
+    assertEquals("FocusMode continuous", call.argument("focusMode"), "auto");
     assertEquals("exposurePointSupported", call.argument("exposurePointSupported"), true);
+    assertEquals("focusPointSupported", call.argument("focusPointSupported"), true);
   }
 
   @Test
