@@ -7,8 +7,11 @@ import 'dart:math';
 
 import 'package:async/async.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
+import 'package:camera_platform_interface/src/events/device_event.dart';
 import 'package:camera_platform_interface/src/method_channel/method_channel_camera.dart';
+import 'package:camera_platform_interface/src/types/focus_mode.dart';
 import 'package:camera_platform_interface/src/utils/utils.dart';
+import 'package:flutter/services.dart' hide DeviceOrientation;
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -25,7 +28,10 @@ void main() {
         MethodChannelMock cameraMockChannel = MethodChannelMock(
             channelName: 'plugins.flutter.io/camera',
             methods: {
-              'create': {'cameraId': 1}
+              'create': {
+                'cameraId': 1,
+                'imageFormatGroup': 'unknown',
+              }
             });
         final camera = MethodChannelCamera();
 
@@ -108,7 +114,10 @@ void main() {
         MethodChannelMock cameraMockChannel = MethodChannelMock(
             channelName: 'plugins.flutter.io/camera',
             methods: {
-              'create': {'cameraId': 1},
+              'create': {
+                'cameraId': 1,
+                'imageFormatGroup': 'unknown',
+              },
               'initialize': null
             });
         final camera = MethodChannelCamera();
@@ -125,6 +134,8 @@ void main() {
           1080,
           ExposureMode.auto,
           true,
+          FocusMode.auto,
+          true,
         ));
         await initializeFuture;
 
@@ -136,6 +147,7 @@ void main() {
             'initialize',
             arguments: {
               'cameraId': 1,
+              'imageFormatGroup': 'unknown',
             },
           ),
         ]);
@@ -162,6 +174,8 @@ void main() {
           1920,
           1080,
           ExposureMode.auto,
+          true,
+          FocusMode.auto,
           true,
         ));
         await initializeFuture;
@@ -205,6 +219,8 @@ void main() {
           1080,
           ExposureMode.auto,
           true,
+          FocusMode.auto,
+          true,
         ));
         await initializeFuture;
       });
@@ -222,8 +238,10 @@ void main() {
           2160,
           ExposureMode.auto,
           true,
+          FocusMode.auto,
+          true,
         );
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('initialized', event.toJson()), cameraId);
 
         // Assert
@@ -242,13 +260,13 @@ void main() {
         // Emit test events
         final fhdEvent = CameraResolutionChangedEvent(cameraId, 1920, 1080);
         final uhdEvent = CameraResolutionChangedEvent(cameraId, 3840, 2160);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', fhdEvent.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', uhdEvent.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', fhdEvent.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('resolution_changed', uhdEvent.toJson()), cameraId);
 
         // Assert
@@ -269,11 +287,11 @@ void main() {
 
         // Emit test events
         final event = CameraClosingEvent(cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('camera_closing', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('camera_closing', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('camera_closing', event.toJson()), cameraId);
 
         // Assert
@@ -292,12 +310,36 @@ void main() {
 
         // Emit test events
         final event = CameraErrorEvent(cameraId, 'Error Description');
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('error', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('error', event.toJson()), cameraId);
-        await camera.handleMethodCall(
+        await camera.handleCameraMethodCall(
             MethodCall('error', event.toJson()), cameraId);
+
+        // Assert
+        expect(await streamQueue.next, event);
+        expect(await streamQueue.next, event);
+        expect(await streamQueue.next, event);
+
+        // Clean up
+        await streamQueue.cancel();
+      });
+
+      test('Should receive device orientation change events', () async {
+        // Act
+        final eventStream = camera.onDeviceOrientationChanged();
+        final streamQueue = StreamQueue(eventStream);
+
+        // Emit test events
+        final event =
+            DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
+        await camera.handleDeviceMethodCall(
+            MethodCall('orientation_changed', event.toJson()));
+        await camera.handleDeviceMethodCall(
+            MethodCall('orientation_changed', event.toJson()));
+        await camera.handleDeviceMethodCall(
+            MethodCall('orientation_changed', event.toJson()));
 
         // Assert
         expect(await streamQueue.next, event);
@@ -332,6 +374,8 @@ void main() {
             1920,
             1080,
             ExposureMode.auto,
+            true,
+            FocusMode.auto,
             true,
           ),
         );
@@ -672,6 +716,54 @@ void main() {
         ]);
       });
 
+      test('Should set the focus mode', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setFocusMode': null},
+        );
+
+        // Act
+        await camera.setFocusMode(cameraId, FocusMode.auto);
+        await camera.setFocusMode(cameraId, FocusMode.locked);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setFocusMode',
+              arguments: {'cameraId': cameraId, 'mode': 'auto'}),
+          isMethodCall('setFocusMode',
+              arguments: {'cameraId': cameraId, 'mode': 'locked'}),
+        ]);
+      });
+
+      test('Should set the exposure point', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setFocusPoint': null},
+        );
+
+        // Act
+        await camera.setFocusPoint(cameraId, Point<double>(0.5, 0.5));
+        await camera.setFocusPoint(cameraId, null);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setFocusPoint', arguments: {
+            'cameraId': cameraId,
+            'x': 0.5,
+            'y': 0.5,
+            'reset': false
+          }),
+          isMethodCall('setFocusPoint', arguments: {
+            'cameraId': cameraId,
+            'x': null,
+            'y': null,
+            'reset': true
+          }),
+        ]);
+      });
+
       test('Should build a texture widget as preview widget', () async {
         // Act
         Widget widget = camera.buildPreview(cameraId);
@@ -685,7 +777,9 @@ void main() {
           () {
         final camera = MethodChannelCamera();
 
-        expect(() => camera.handleMethodCall(MethodCall('unknown_method'), 1),
+        expect(
+            () =>
+                camera.handleCameraMethodCall(MethodCall('unknown_method'), 1),
             throwsA(isA<MissingPluginException>()));
       });
 
@@ -765,6 +859,41 @@ void main() {
                 .having((e) => e.code, 'code', 'ZOOM_ERROR')
                 .having((e) => e.description, 'description',
                     'Illegal zoom error')));
+      });
+
+      test('Should lock the capture orientation', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'lockCaptureOrientation': null},
+        );
+
+        // Act
+        await camera.lockCaptureOrientation(
+            cameraId, DeviceOrientation.portraitUp);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('lockCaptureOrientation',
+              arguments: {'cameraId': cameraId, 'orientation': 'portraitUp'}),
+        ]);
+      });
+
+      test('Should unlock the capture orientation', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'unlockCaptureOrientation': null},
+        );
+
+        // Act
+        await camera.unlockCaptureOrientation(cameraId);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('unlockCaptureOrientation',
+              arguments: {'cameraId': cameraId}),
+        ]);
       });
     });
   });
